Gjort om utils til service klasser
